pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "envctl"
version = "2.0.0-dev"
description = "Local environment control plane for contract-driven development workflows"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Alessandro" }
]
dependencies = [
  "typer>=0.12,<1.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=7.1.0,<8.0.0",
  "ruff>=0.11,<0.12",
]

[project.scripts]
envctl = "envctl.cli.app:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
